Move attributes model into shader module
